Update Seqeunce_Check.R
This file confirms single amino acid modifications that I use to verify information in the Leung Lab database. Three files are required: text file listing the Uniprot IDs for proteins of interest, a text file listing the respective modifications for each protein of interest, and a FASTA file containing information on each protein of interest. Proteins for which the reported modifications are incorrect are returned.