Update cleaning_Domain_text.R
This file summarizes and reports protein domain information given a text file downloaded from the Uniprot website. The name of the domain feature, its description, and amino acid positions are reported. This code was created to prepare domain data to be used in a protein graphic for the Leung Lab protein database.